[REF] sale_restricted_qty: Merge duplicated model inheritance
Merge duplicated 'sale.order.line' model inheritance from sale.py and sale_order_line.py

The improved model structure from sale_order_line.py replaces the original migration model.

This resolves pylint warning R8180 about considering merging classes inherited to the same model.
