+++ sale_restricted_qty/models/__init__.py
@@ -1,2 +1,2 @@
 from . import product_restricted_qty_mixin  # noqa
-from . import product_category, product_product, product_template, sale_order_line
+from . import product_category, product_product, product_template, sale
